bots/localizei/requirements.txt
Flask==2.2.5
requests==2.31.0
urllib3==1.26.14
python-dotenv==1.0.0
2captcha-python==1.2.8
lxml==5.2.2
